campaign2: main recovered (310->315), and with it five helpers
main (0x5546) is byte-identical.  It moves into stx_u1 and INLINES
five things the port had factored out, which is the whole reason it
is 1370 bytes against the port's 602:

 * ct_clrB -- the supervisor-mode conterm patch.
 * The .SCN file handling.  Only the nibble decoder is a function
   (parts/scn_dec.c, 0x52ca); the size, buffer and 15-word dictionary
   are globals, and the file handle is never closed.
 * al_locs -- body.lcp loads FIRST, then lcp_crnd for a new game,
   then the PEx filename is patched and loaded.  al_loal (0x6428)
   takes only (name, buffer): no size cap, no return value, and the
   header's second word IS the length.
 * al_loot/al_lost -- fixed 56- and 50-iteration walks with no
   zero-record or size checks.
 * sp_reglp -- its loop, calling sp_regs (0x5bdc) per sprite.

Also: STX calls `Dsetpath("data")` (the port's comment claimed there
was none); cp_main returns a LONG; the door draws are seven full
if/else pairs rather than ternaries in the argument, with only the
cabinet testing `== NO`; the dog bowl is three explicit state tests
with literal frame ids; and main ends at gameLoop with no Pterm.

sp_regs reads its four values back out of the tables instead of using
its parameters; sp_genma (0x408c) keeps the words-per-row division in
a local and walks with *p++.  pex_name is a `char *`, not an array.

al_locs, al_loot, al_lost, unScn, prsRec, sp_reglp and ct_clrB are
deleted -- they do not exist in LCP_STX.

+++ source/include/stubs.h
@@ -3,6 +3,6 @@
 #ifndef STUBS_H
 #define STUBS_H
 
-extern short cp_main();
+extern long cp_main();
 
 #endif /* STUBS_H */
